backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["YingShen <yingshen730@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
langchain = "^0.2.5"
qdrant-client = "^1.9.1"
python-decouple = "^3.8"
langchain-community = "^0.2.5"
openai = "^1.35.1"
langchain-openai = "^0.1.8"
bs4 = "^0.0.2"
langchain-qdrant = "^0.1.1"
pypdf = "^4.2.0"
langchain-groq = "^0.1.5"
uuid = "^1.30"
boto3 = "^1.34.131"
authlib = "^1.3.1"
starlette = "^0.37.2"
itsdangerous = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
unstructured = "^0.14.8"
modal = "^0.63.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
